# kv: replace memdb with a more memory efficient version