@ngaut 2019-08-26
/run-all-tests