@ngaut 2019-08-28
/run-all-tests